Replace stub with mock
