+++ default/python/tests/conftest.py
@@ -5,18 +5,28 @@ https://docs.pytest.org/en/2.7.3/plugins.html?highlight=re#working-with-plugins-
 
 import os
 import sys
+from unittest.mock import MagicMock
 
 this_dir = os.path.dirname(__file__)
-
-sys.path.append(this_dir)
 sys.path.append(os.path.join(this_dir, "..", "..", "python", "AI/"))
 sys.path.append(os.path.join(this_dir, "..", "..", "python"))
 
-# Since the 'true' freeOrionAIInterface is not available during testing, this import loads the 'mock' freeOrionInterface
-# module into memory.  For testing purposes the mock interface provides values for certain constants which various
-# AI modules rely upon.  Since any later imports of the freeOrionInterface module will simply refer back the first-
-# loaded module of the same name, code from the AI modules can then be imported and tested, provided that such
-# AI code only relies upon constants from the freeOrionAIInterface (and specifically does not depend on any of the
-# various dynamic galaxy-state queries provided by the 'true' freeOrionAIInterface).
-# noinspection PyUnresolvedReferences
-import freeOrionAIInterface
+
+class aggression:
+    beginner = 0
+    turtle = 1
+    cautious = 2
+    typical = 3
+    aggressive = 4
+    maniacal = 5
+
+
+def userString(x):
+    return f"UserString {x}"
+
+
+fo = MagicMock(name="freeOrionAIInterface")
+fo.aggression = aggression
+fo.userString = userString
+
+sys.modules["freeOrionAIInterface"] = fo
